refactor: Move affective inputs to affective annotation project
